fastmvc_storage/pyproject.toml
[project]
name = "fastmvc_storage"
version = "0.1.0"
description = "Object storage (S3, GCS, Azure Blob) for FastMVC."
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastmvc_core",
]

[project.optional-dependencies]
dev = ["pytest>=7.0.0", "pytest-cov>=4.0.0"]
s3 = ["boto3>=1.28.0"]
gcs = ["google-cloud-storage>=2.10.0"]
azure = ["azure-storage-blob>=12.17.0"]

[project.urls]
Homepage = "https://github.com/your-org/fastmvc_storage"

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
where = ["src"]
